Update hard shutdown, raise shutdown error on threads
This updates the hard shutdown logic to force worker threads down
using `Thread#raise`. This is not ideal but there isn't any other
way to force threads down if they are processing work. The hard
shutdown was and still is a "use at your own risk" feature. This
is slightly better handling because threads will be shutdown by
the raised exception and not by ruby (which uses `Thread#kill` and
is known to cause problems). It will also give error handling in
do work procs to have a chance to handle the error. This is part
of a set of changes to have better hard-shutdown logic to help
avoiding losing work.