Add custom error handling to workers
This updates workers to allow passing a proc for custom error
handling. This is needed because we are now using `Thread#raise`
to kill workers threads when forcing a worker pool to shutdown.
Since these exceptions can be raised at any point, we need to
allow systems to provide custom error handling when this occurs.
This expands a workers error catching to catch both standard errors
and hard shutdown errors for its entire loop. For hard shutdown
errors, after custom error handling is called, the worker will
reraise the error which ensures that thread will exit from the
exception.

This also fixes a small bug where it was possible that work
could be pulled off of the queue but could be shutdown before
processing it. This would result in the work being lost. Now if
work is pulled off, it will attempt to process it always. If the
hard shutdown raises at this point then the custom error handling
has a chance to deal with the work.