Consistently organize methods, cleanup comments, simplify timeout
This is a set of cleanups that doesn't change any functionality.
These changes are part of keeping the code easier to maintain and
understand. It has gone through a few rounds of fairly significant
changes so this seeks to cleanup the code that has been added and
changed.

It organizes methods so that they are more consistent.
Specifically, this moves the start/shutdown methods to the top for
the worker pool, queue and workers. This also simplifies some of
the comments explaining some of the more strange or confusing parts
of the code.

Finally, this simplifies the shutdown timeout handling. The
optional timeout has been simplified and the re-raise logic
removed. It now also raises a timeout error which then triggers
the worker pool to raise a shutdown error. This makes more sense
than the timeout raising a shutdown error directly because the
timeout triggers the hard shutdown (aka the shutdown error).