Add on worker error callback to `WorkerPoolSpy`
This updates the worker pool spy to have the same interface as the
real worker pool. This adds the on worker error callback methods
to it. This should've been done when they were added to the main
worker pool but was accidentally forgotten. This ensures the spy
is still a good stand-in for a real pool.

This also does some minor cleanups to the spy to keeps its methods
ordered similarly to the real pool. This should've also been done
when the real pools methods order was updated but it was also
forgotten.