Remove min and max workers, always spawn a consistent number of workers
This changes worker pools to alway spawn a set number of workers.
They no longer have a min and max setting. This simplifies the
worker pools logic and is setup for allowing custom queue classes.

Previously, dat worker pool would allow specifying a min and max
number of workers. When work was added it would see if a worker
was available to process it, and if not, spawn another worker if
it hadn't reached the max. With allowing custom queue classes, this
option no longer makes as much sense. For some custom queue classes
they won't define a `push` method because they will not have work
directly pushed onto them. Instead they will get their work from
redis or a socket. In this case, there is no logical place to
trigger spawning additional workers. If we left min and max number
of workers in, you would have to know to always configure the min
and max the same if you had a queue that couldn't take advantage
of it. Rather than deal with this and the added confusion, its
simpler to remove the logic.