Fix infinite loop bug when force shutting down
This fixes a rare but possible infinite loop bug when force
shutting down a worker pool.

Force shutting down raises shutdown errors in the worker pools
worker threads. This causes them to error and exit out. Its
possible that the worker thread is in a state when the shutdown
error is raised it doesn't catch it (the `ensure` block in its
`work_loop` method for example.). If this happens, the previous
logic would catch and ignore the error but wouldn't remove the
worker from its collection. Its `until` loop would iterate and
pick the same worker and infinitely repeat. This fixes it by
manually deleting the worker from the collection at the end of
the loop (when it normally shuts down it deletes itself from the
collection).

This also adds tests for this behavior and changes the forced
shutdown tests. They now use fake/spy workers to setup the specific
scenarios instead of trying to create the scenario by manipulating
actual threads. This is part of an effort to make dat worker pools
unit tests cleaner and less "systemy". As part of this, the
previous forced shutdown tests were moved into the system tests
(because they manipulate threads and try to craft specific
scenarios).

Finally, this also removes "worker pool" from the shutdown error
raised when forcing the worker pool to shutdown. This error can
be caught and reported on in systems built on dat worker pool and
I'd rather the message not reveal anything about dat worker pool.