Rename `Queue` to `DefaultQueue`
This renames the `Queue` class to `DefaultQueue`. This is setup for
having a base `Queue` mixin that can be used to define custom
queues for dat worker pool. This renames the existing queue as
a first step so the `Queue` name is now available. The
`DefaultQueue` will continue to be provided by dat worker pool and
will also continue to be the default for a worker pool.