Add `Queue` mixin for defining custom queues
This adds a `Queue` mixin which defines the interface needed to
define a custom queue for dat-worker-pool. This also updates the
`DefaultQueue` to use it.

The `Queue` mixin defines the interface that dat-worker-pool
expects its queues to have. It is intended to be mixed into a
queue class and have its `!` methods overwritten. It provides some
default logic for handling starting and shutting down the queue
but doesn't provide any "queue" logic. A custom queue class is
responsible for the "queue" logic and also making it thread-safe.

With the new mixin, queues now use a running flag instead of a
shutdown flag. The biggest difference with this change is that
the default state of a queue is not running or shutdown. This is
more correct, a queue `start` method should be called before using
the queue, but caused some of the tests to have to start queues
before using them.

This also updates the `DefaultQueue` to use the `Queue` mixin and
also be an example of how to define a custom queue. This mostly
involved removing duplicate logic that the `Queue` now provides
and updating the default queues logic to use the queue mixin
interface. This doesn't change the functionality of the default
queue just how its defined.