Change `DefaultQueue` callbacks, pass the queue and work item
This updates the `DefaultQueue` callbacks to pass the queue and
work item to them. This makes them consistent with the worker
callbacks and makes them more useful. Now they can use the queue
or work item if needed.

This also fixes a bug with running the on pop callbacks even if
they popped a `nil` work item. Dat worker pool as a whole ignores
`nil` work items and this makes the on pop consistent with that
behavior. The bug this fixes is running on pop callbacks when a
queue is shutdown with a worker thread waiting on pop. Prevously
the worker thread would get a `nil` work item but it would have
caused the on pop callbacks to run.