Add `Runner` class and `Worker` mixin
This adds a `Runner` class and `Worker` mixin. This is part of an
effort to allow providing a custom worker class and in general make
dat worker pool more flexible and powerful.

This adds a `Runner` which is built by a worker pool and passed
to a worker. This is so we can remove the default worker callbacks
that the worker pool added to its workers for despawning workers
and incrementing/decrementing waiting worker counts. Instead of
using the callbacks and relying on ruby's closures to allow
calling private methods on the worker pool, the worker now
directly calls the runner. When the worker sleeps/wakes up it
calls a public increment/decrement method on the runner. When
it shuts down it calls the public despawn worker method on the
runner. The runner allows methods that are intended to be private
and internal to the working of the worker pool to be public so
the worker can call them.

As part of adding a runner, the start and shutdown logic was
moved into the runner. This allows the runner to control the
order the queue and workers are started the same way the worker
pool used to. Otherwise, the runner has to provide extra methods
for shutting down the workers and waiting on them to finish so
the worker pool can control the order.

This also adds a `Worker` mixin that will eventually be the base
mixin that dat worker pool provides for defining custom workers.
For now, this moves the callbacks logic into it as class DSL
methods. This is a friendlier interface than the previous instance
methods. This also mixes in the `Worker` mixin into the default
worker.

As part of changing the callbacks, this updates how a worker pool
proxies the callback methods. The callbacks are stored on the class
not the instance of the default worker. The dat worker pool
callback methods functionally work the same but they access its
worker class DSL methods. This will be removed once a worker class
can be passed to a worker pool as using the worker class DSL is
the preferred method to add callbacks.