Add the rest of the `Worker` logic from `DefaultWorker`
This adds the rest of the logic from `DefaultWorker` to the
`Worker` mixin. This fills out the worker with all the logic
needed to define custom workers.

The `Worker` mixin now contains all the logic needed to define
custom workers. It knows how to start, shutdown and process work
items while running. Most of the logic was copied straight out of
the default worker but a few things were changed:

* The mixin now expects a `work!` method to be defined. This is
  tells it how to process a work item. It doesn't accept an on
  work proc.
* Switch from a shutdown flag to a running flag. This is consistent
  with queues and makes the default state of the worker not running
  (shutdown).
* This changes the `running?` method to be based on the running
  flag instead of if the thread is alive. This is for consistency
  with queues. This adds a `thread_alive?` to check if the thread
  is alive. This gives us the ability to ask if the worker has been
  shutdown even if its thread is still alive (meaning its in the
  process of shutting down).

This also redoes the unit tests for the logic that was moved to the
`Worker` mixin. They have been improved with more detailed and
granular testing of the worker behavior. They also use our latest
testing conventions.

Finally, this removes most of the logic from `DefaultWorker` only
leaving its `on_work` proc. It now uses the `work!` interface that
the `Worker` mixin expects and calls its `on_work` proc.