Allow passing a worker class to a worker pool
This updates dat-worker-pool to allow passing a worker class. This
is the final part of allowing custom worker classes for defining
how a worker pool processes work.

A worker class can now be passed to a dat-worker-pool. The worker
pool will validate that its a class and that it uses the `Worker`
mixin. This is to ensure the worker class has the correct interface
for working with the worker pool. This also updates the
dat-worker-pool to allow passing worker params which will be passed
down to worker instances. This allows passing in values to a worker
that can be used to process work items.

This also removes all the callback methods on dat-worker-pool. Now
that a worker class can be passed in its class DSL methods should
be used instead.