Change worker callbacks, instance exec and begin-rescue them
This changes the worker callbacks to be instance exec'd, wraps
the on-start and on-shutdown in begin-rescues and changes the
worker to spawn itself with the runner when its started. This is
to make the callbacks more consistent and safer (keep them from
messing up the worker behavior).

The worker callbacks are now instance exec'd instead of being
called. This makes them more consistent with the `work!` method
and lets them easily access the queue and params.  As part of this,
the internal methods and ivars for the workers behavior have been
namespaced to avoid user defined methods and ivars.

This also wraps the on-start and on-shutdown callbacks in
begin-rescues. This is to ensure that if they raise an error, the
on-error callbacks get called. This also tries to keep the errors
from keeping workers from starting or shutting down incorrectly.

Finally, this changes the worker to call a `spawn_worker` method
on the runner when it starts. This is consistent with how it calls
`despawn_worker` when it shuts down. The `spawn_worker` adds the
worker to the runners workers just like the `despawn_worker`
removes it. This makes sure we don't add the worker to the
collection before the thread has started and gotten to run (meaning
its actually processing work).