Changes to shutdown logic
This is a number of small changes to the shutdown logic. In general
these are related to error handling when shutting down. Now that
its possible for queues and workers to be customized its possible
for more errors to be introduced. This ensures we cover those
scenarios and try to report the errors as much as possible.

* The workers and queue are now signalled to shutdown outside the
timeout so they will always be "shutdown" even if the timeout is
short.

* Runner shutdown will now rescue errors when joining workers and
if an error occurs it will manually remove it. This ensures that if
a worker errors it doesn't get the other workers killed when they
could shutdown normally.

* Runner shutdown will now catch standard errors that are thrown
and do a forced shutdown. This avoids the process exiting and
the workers getting killed. The error will be re-raised after the
runner has finished shutting down.

* The `TimoutError` is now named `TimeoutInterruptError` and is
not a standard error. The timeout error is typically thrown from
the `worker.join` which is now rescued. This caused the shutdown to
no longer timeout. This error is like the shutdown error, its
internal and shouldn't be rescued.

* Fixes the forced shutdown error backtrace to once again be
the caller of the dat-worker-pool `shutdown` method. This keeps
the backtrace out of the internals of the worker and makes it
relevant to what action caused the error. This was what it was set
to before we added a `Runner`.

* Fixes the runner shutdown tests to ensure they actually test
workers. Previously they would rely on the runners workers array
to do assertions but this array is not safe because it is emptied
as the workers shutdown. The tests now get a copy of the workers
array and use that to do their assertions.