Update `WorkerPoolSpy` and add `Worker::TestHelpers`
This updates the worker pool spy with all the changes that were
made to the worker pool. This also adds worker test helpers for
help testing custom workers.

The worker pool spy is a test helper that can be stubbed in as a
replacement to a real worker pool. This updates it so it continues
to provide the same interface as a real worker pool. This keeps it
relevant and useful as a test helper.

This also adds worker test helpers that help with building a worker
and running its behavior. The test runner it builds can be used to
run the workers callbacks and work method separately or all
together. This should help with testing custom worker logic and
making sure it works as expected.