Add `LockedObject`, `LockedArray` and `LockedSet` utility classes
This adds `LockedObject`, `LockedArray` and `LockedSet` utility
classes which commonizes an object and mutex being used together.
The mutex is used to lock interactions which the object so it makes
sure only one thread is operating on it at a time. This was already
being done through most of dat-worker-pool, this just formalizes it
and also provides it as a helper for other gems that use
dat-worker-pool.

This adds a base `LockedObject` class which takes a value and
builds a mutex for it. The mutex is used to lock access to the
value. The base `LockedObject` only provides a method for reading
the value, a method for setting it and a helper for accessing the
value with the lock. This is used as the base for the `LockedArray`
and `LockedSet` class.

The `LockedArray` class takes an array and demeters it while using
the mutex to lock access. This is used by the runner for its
workers and the default queue for its work items. By using this in
the runner it fixes a rare bug that could happen when joining
workers while shutting down. Because accessing the workers array
wasn't locked with a mutex, a worker could remove itself while we
were asking for the first worker. If it was the last worker, this
caused `nil` errors. The work items was already locked by a mutex
that the default queue uses to signal workers but this ensures that
when adding additional logic it will lock access as well.

The `LockedSet` takes a set and demeters it similarly to the locked
array. It is used by the runner to track available workers. This
was already being done manually by the runner but it now uses the
utility class instead of its own custom class.