Don't have workers add/remove themselves to/from the runner
This changes workers to no longer add themselves to or remove
themselves from the runners workers array. This fixes issues with
starting and shutting down a runner before workers have gotten a
chance to add themselves to the runners workers array. If this
happens it causes the workers to run endlessly and never get shut
down.

This runner will now add workers when it builds them. This means
workers will always get added. Before it was possible the runner
could get shutdown before the workers had a chance to add
themselves to the runner. This ensures the runner always has
access to the workers to shut them down.

The runner now also always removes its workers after it joins them
during a normal shutdown. Previously, it would only remove the
workers if they errored. It always removes them when it forces them
to shutdown like it previously did. Both the normal and forced
shutdown will also remove workers from the available set in case
the workers were not able to do this themselves.

The workers no longer add or remove themselves now, but they will
mark themselves available and unavailable when started and shut
down. This was previously being done when they were added or
removed which was an issue because we were making them available
or unavailable without calling their callbacks. Now they will
properly call the callbacks when they make themselves available.