Rework system tests
This reworks the system tests to be more maintanable, match our
latest conventions and faster.

The tests no longer rely on sleeping or "polling" to achieve
certain scenarios. Instead they use mutexes and condition variables
to ensure certain scenarios are setup. These allow the workers and
test suite thread to signal one another and ensure the workers are
in a specific state before we make an assertion.

This also switches the tests to use randomized data as much as
possible which ensures we don't have magic values that conveniently
make the tests pass.

Finally, this adds a `size` method to the `LockedArray` which is
used by the new system tests.