Update workers to signal their test suite thread in worker tests
This updates the workers used in the worker tests to signal their
test suite thread as they move into different states. This makes
the test less fragile and ensures the worker has reached the state
we expect before try and make any assertions. This mimics changes
that were done to the dat-worker-pool system tests to make them
less fragile.

Specifically, this fixes an issue when logging is added to the
workers. This slows down the workers and causes ruby to switch to
other threads. This means it can switch to the test suite thread
and make assertions before the worker has had a chance to do
anything. This causes tests to fail even though they shouldn't if
the worker thread has a chance to run.

This also includes some cleanups to the tests to try and
commonize as much of their setup as possible. This involved
expanding the thread spy to actually run a thread and moving it
into the thread spies support file.

This also fixes a rare related issue in the default queue tests.
Sometimes the thread used in the tests doesn't get to run so the
tests randomly fail. This ensures that it gets to run by joining
it for a short period.