Update worker to handle errors before worker is made available
This updates the worker to run its on-error callbacks when an error
occurs before making itself available. If an error happened either
while making itself unavailable or while working, it would
previously make itself available before handling the error. This
is not the correct order these events should run in, the error
should be handled before the worker is made available.

This also addresses a situation where the worker could be crashed
if the on-error callbacks raised an error. Now those errors will
be ignored and the worker will continue to run. The danger of this
is that a worker could be erroring and its on-error callbacks could
be erroring. This would essentially hide that there was a problem.
Ideally, on-error callbacks should catch their own errors and
handle them however is best for the user (log, send an email, etc).