Add `work_items` to pool and queue, make pool spy use queue
This updates `DatWorkerPool` and `Queue` to have `work_items`
methods and updates the `WorkerPoolSpy` to use its queue for its
work items. This makes using the default queue with a worker pool
simpler and makes the worker pool spy match the real worker pool
more making it a better spy.

Gems (specifically, dat-tcp) need to read the current work items
on the worker pool's default queue. Without a `work_items` method
on the worker pool they have to manually build a default queue
and pass that into the worker pool. This makes it more annoying to
use a default queue. Ideally, if you want to use the default queue
you never need to pass it into the worker pool.

To allow `DatWorkerPool` to provide a `work_items` method the
`Queue` mixin now provides it by default but it raises a not
implemented error. Like the other methods it provides, it can
be re-defined to provide custom logic. The `DefaultQueue` already
does this allowing access to its in-memory array of work items.

This also changes the worker pool spy to use its queue. Previously,
it took a queue but ignored it and stored its work items on itself.
This doesn't work as expected if a custom queue is passed in, it
wouldn't add work items to the custom queue like a real worker pool
would. To make sure it behaves more like a real worker pool, it
will now start, shutdown, add work and read work items from its
queue just like the real worker pool does.

This also adds some missing spy methods to the worker pool spy.