Add logging to runner and workers
This updates the runner and workers to have debug logging. This is
to help with seeing what a worker pool and its workers are doing.

The runner will now log when it starts and shuts down. It logs the
number of workers it is starting or waiting to shut down and will
log if it is doing a normal shutdown or forcing the workers to
shutdown.

The worker will log as it moves between its different states:
starting, available, unavailable, working and shutting down. It
also logs any errors that occur.

In addition to adding logging, workers are now given a number by
their runner. This is used in their logs to identify the worker
and is also a helper that can be used in custom logging or other
logic as needed.