Update test suite to work with newer versions of ruby
This updates the test suite to work with newer versions of ruby.
This mostly involved doing more "hand-holding" to ensure threads
were given a chance to run or to reach a certain state before any
assertions were done. This gets all the tests passing in newer
versions of ruby and in ruby 1.8.7.

This fixes the default queue unit tests by simply adding some extra
thread joins to ensure it gets a chance to run. This allows it to
reach the expected state before any assertions happen.

This fixes the runner unit tests by making a few changes:

* The workers now signal when they become available and
  unavailable. This allows the tests to ensure the workers have
  reached an expected state before running any assertions. This
  logic was commonized from the system tests so both sets of tests
  can make use of it.
* The workers no longer override their `dwp_raise` method which
  caused them to not behave as they normally would. Instead, they
  spy what errors occurred by using the `on_error` callback.
* The workers no longer raise a join error via their `dwp_join`
  method. They now use the `on_shutdown` callback to raise join
  errors which follows the normal behavior of a thread raising
  any unhandled errors when its joined.
* The workers now suspend their state to test how they can be
  forced to shutdown by running special "hang" work items.
  Previously this was done during shutdown and via the overriden
  methods. Now that the methods aren't overriden the old logic
  didn't work.

All of these changes fixed the runner tests in newer versions of
ruby. Overall, these changes let the workers behave normally which
avoids issues between the different versions of ruby.

This also cleans up a few spots that were doing special behavior
because they expected system timer and its aggressive deadlock
catching. This is no longer needed now that we are using
`MuchTimeout`. These were locations that system timer thought
were deadlocks but weren't because we were running them in a
timeout.

Finally, this includes some minor cleanups to switch to using
`MuchTimeout` instead of `Timeout`. This is more consistent and
ensures we don't have issues if `Timeout`'s implementation changes
between ruby versions. As part of this, the worker unit tests were
updated to only stub `Thread.new` the first time it was called
because `MuchTimeout` would call `Thread.new` and use the stub
which caused errors. This makes the stub better and only stub its
`Thread.new` call and not other unintended calls.