Rescue `Timeout::Error` in `Worker` and handle like a standard error
This updates the `Worker` to rescue `Timeout::Error` and treat it
as a standard error. This is to fix an issue where timeout errors
were causing worker threads to shutdown. This is because timeout
errors don't inherit from `StandardError` so they wouldn't be
caught and the worker's work loop would exit. Timeout errors are
now rescued with standard errors. This means they will run through
the error callbacks for the worker and then be ignored which keeps
the work loop from exiting.