Rescue `LoadError` and `NotImplementedError` in `Worker`
This updates the `Worker` to rescue `LoadError` and
`NotImplementedError` exceptions. These are non `StandardError`
exceptions that would cause the worker thread to exit if they
aren't rescued. It's relatively likely these can occur in code
being run by a worker. It's a better workflow for these to be
rescued and passed to any configured error procs rather than a
worker shutting down. This allows the error procs to notify that
there was a problem. Otherwise the worker thread exits and there
isn't any chance for a notification.